動的に画面を切り替える処理を追加

+++ src/main/java/com/example/demo/damain/repository/ClothesRepository.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ClothesRepository extends JpaRepository<Clothes, Integer> {
+public interface ClothesRepository extends JpaRepository<Clothes,Long> {
     List<Clothes> findAllByUserId(Long userId);
 
     List<Clothes> findAllByUserIdOrderByCreatedAtAsc(Long userId);
@@ -19,4 +19,5 @@ public interface ClothesRepository extends JpaRepository<Clothes, Integer> {
     List<Clothes> findAllByUserIdOrderByCreatedAtDesc(Long userId);
 
     List<Clothes> findAllByUserIdAndStorageId(Long userId, Long storageId);
+
 }
